+++ packages/effect-app/src/client/errors.ts
@@ -108,10 +108,11 @@ export class OptimisticConcurrencyException extends TaggedError<OptimisticConcur
   { message: S.String }
 ) {
   readonly details?: OptimisticConcurrencyDetails
+  readonly raw?: unknown
   constructor(
     args:
       | OptimisticConcurrencyDetails
-      | (S.Struct.Constructor<typeof OptimisticConcurrencyException.fields> & { cause?: unknown }),
+      | (S.Struct.Constructor<typeof OptimisticConcurrencyException.fields> & { cause?: unknown; raw?: unknown }),
     disableValidation?: boolean
   ) {
     super("message" in args ? args : { message: `Existing ${args.type} ${args.id} record changed` }, disableValidation)