+++ packages/infra/src/Model/filter/filterApi.ts
@@ -1,3 +1,5 @@
+import { type RelationDirection } from "../query.js"
+
 export type InOps =
   | "in"
   | "notIn"
@@ -26,13 +28,16 @@ export type Ops = OtherOps | InOps
 
 export type FilterScopes = {
   t: "or-scope"
-  result: FilterResult[]
+  result: readonly FilterResult[]
+  relation: RelationDirection
 } | {
   t: "and-scope"
-  result: FilterResult[]
+  result: readonly FilterResult[]
+  relation: RelationDirection
 } | {
   t: "where-scope"
-  result: FilterResult[]
+  result: readonly FilterResult[]
+  relation: RelationDirection
 }
 
 export type FilterR = {