+++ packages/infra/src/Model/query/dsl.ts
@@ -52,17 +52,21 @@ type ExtractFieldValuesRefined<T> = T extends QueryTogether<any, infer TFieldVal
   ? TFieldValuesRefined
   : never
 
+export type RelationDirection = "some" | "every"
+export type Relation = { relation: RelationDirection }
 export type Query<TFieldValues extends FieldValues> = QueryTogether<TFieldValues, TFieldValues, false, "initial">
 export type QueryWhere<
   TFieldValues extends FieldValues,
   TFieldValuesRefined extends TFieldValues = TFieldValues,
   Exclusive extends boolean = false
-> = QueryTogether<
-  TFieldValues,
-  TFieldValuesRefined,
-  Exclusive,
-  "where"
->
+> =
+  & QueryTogether<
+    TFieldValues,
+    TFieldValuesRefined,
+    Exclusive,
+    "where"
+  >
+  & Relation
 
 export type QueryEnd<
   TFieldValues extends FieldValues,
@@ -120,6 +124,8 @@ export class Initial<TFieldValues extends FieldValues> extends Data.TaggedClass(
 export class Where<TFieldValues extends FieldValues> extends Data.TaggedClass("where")<{
   current: Query<TFieldValues>
   operation: [string, Ops, any] | [string, any] | ((q: Query<TFieldValues>) => QueryWhere<TFieldValues>)
+  relation: RelationDirection
+  subPath?: string
 }> implements QueryWhere<TFieldValues> {
   readonly [QId]!: any
 
@@ -132,6 +138,7 @@ export class Where<TFieldValues extends FieldValues> extends Data.TaggedClass("w
 export class And<TFieldValues extends FieldValues> extends Data.TaggedClass("and")<{
   current: Query<TFieldValues>
   operation: [string, Ops, any] | [string, any] | ((q: Query<TFieldValues>) => QueryWhere<TFieldValues>)
+  relation: RelationDirection
 }> implements QueryWhere<TFieldValues> {
   readonly [QId]!: any
   pipe() {
@@ -143,6 +150,7 @@ export class And<TFieldValues extends FieldValues> extends Data.TaggedClass("and
 export class Or<TFieldValues extends FieldValues> extends Data.TaggedClass("or")<{
   current: Query<TFieldValues>
   operation: [string, Ops, any] | [string, any] | ((q: Query<TFieldValues>) => QueryWhere<TFieldValues>)
+  relation: RelationDirection
 }> implements QueryWhere<TFieldValues> {
   readonly [QId]!: any
   pipe() {
@@ -223,6 +231,29 @@ export const and: FilterContinuationAnd = (...operation: any[]) => (current: any
 
 export const or: FilterContinuationOr = (...operation: any[]) => (current: any) =>
   new Or({ current, operation: typeof operation[0] === "function" ? flow(...operation as [any]) : operation } as any)
+
+// TODO: make nice.
+export type WhereEveryOrSome = {
+  <
+    TFieldValues extends FieldValues,
+    TFieldName extends FieldPath<TFieldValues>
+  >(
+    path: TFieldName,
+    dude: (
+      current: Query<TFieldValues[TFieldName][number]>
+    ) => QueryWhere<TFieldValues[TFieldName][number], TFieldValues[TFieldName][number], false>,
+    ...dudes: ((
+      current: QueryWhere<TFieldValues[TFieldName][number], TFieldValues[TFieldName][number], false>
+    ) => QueryWhere<TFieldValues[TFieldName][number], TFieldValues[TFieldName][number], false>)[]
+  ): (
+    current: Query<TFieldValues>
+  ) => QueryWhere<TFieldValues, TFieldValues, false>
+}
+
+export const whereEvery: WhereEveryOrSome = (subPath, ...operations) => (current) =>
+  new Where({ current, operation: flow(...operations as [any]), relation: "every", subPath } as any)
+export const whereSome: WhereEveryOrSome = (subPath, ...operations) => (current) =>
+  new Where({ current, operation: flow(...operations as [any]), relation: "some", subPath } as any)
 
 export const order: {
   <
@@ -465,7 +496,7 @@ export type RefineWithLiteral<
   Exclde extends boolean = false
 > =
   // refine only if the value is a primitive
-  [V] extends [string | number | boolean | null | bigint | undefined]
+  [V] extends [string | number | boolean | null | bigint]
     ? RefineFieldPathValue<TFieldValues, TFieldName, V, Exclde>
     : TFieldValues
 /* dprint-ignore-end */