+++ packages/vue-components/stories/OmegaForm.stories.ts
@@ -15,6 +15,9 @@ import FormInputComponent from "./OmegaForm/form.Input.vue"
 import OneHundredWaysToWriteAFormComponent from "./OmegaForm/OneHundredWaysToWriteAForm.vue"
 import ClearableComponent from "./OmegaForm/Clearable.vue"
 import BooleansComponent from "./OmegaForm/Booleans.vue"
+import DateComponent from "./OmegaForm/Date.vue"
+import ArrayComponent from "./OmegaForm/Array.vue"
+import TanstackComponent from "./OmegaForm/Tanstack.vue"
 
 const mockIntl = {
   locale: ref("en"),
@@ -126,5 +129,25 @@ export const Booleans: Story = {
   render: () => ({
     components: { BooleansComponent },
     template: "<BooleansComponent />",
+  }),
+}
+
+export const Date: Story = {
+  render: () => ({
+    components: { DateComponent },
+    template: "<DateComponent />",
+  }),
+}
+export const Array: Story = {
+  render: () => ({
+    components: { ArrayComponent },
+    template: "<ArrayComponent />",
+  }),
+}
+
+export const Tanstack: Story = {
+  render: () => ({
+    components: { TanstackComponent },
+    template: "<TanstackComponent />",
   }),
 }